Decompose Codex JSONL scanner
